requirements.txt
alabaster==0.7.12
amqp==5.0.6
appdirs==1.4.4
astor==0.8.1
astroid==2.4.2
attrs==20.3.0
autoflake==1.4
autopep8==1.5.6
Babel==2.9.0
backoff==1.11.1
bandit==1.7.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==20.8b1
branca==0.4.2
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.1
coverage==5.4
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
darglint==1.7.0
decorator==4.4.2
distlib==0.3.1
doc8==0.8.1
docopt==0.6.2
docutils==0.16
docx==0.2.4
eradicate==1.0
feedparser==6.0.2
filelock==3.0.12
fire==0.4.0
flake8==3.9.0
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
folium==0.12.1
fuzzywuzzy==0.18.0
gdown==3.12.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.13
google==3.0.0
greenlet==1.1.2
gunicorn==20.0.4
highlight-text==0.1
ibm-cloud-sdk-core==1.7.3
ibm-watson==5.0.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imagesize==1.2.0
inflect==5.0.2
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
jaraco.classes==3.2.0
jaraco.collections==3.1.0
jaraco.functools==3.1.0
jaraco.text==3.4.0
Jinja2==2.11.2
joblib==1.0.0
Js2Py==0.70
kiwisolver==1.3.1
kombu==5.0.2
lazy-object-proxy==1.4.3
lxml==4.6.2
map==1.3.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==8.6.0
moviepy==1.0.3
munch==2.5.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
networkx==2.5
nltk==3.5
num2words==0.5.10
numpy==1.19.4
opencage==2.0.0
packaging==20.9
pandas==1.2.3
parse==1.19.0
pathspec==0.8.1
pattern3==3.0.0
pbr==5.5.1
pdfminer.six==20201018
pdfminer3k==1.3.4
pep8-naming==0.9.1
phonenumbers==8.12.31
Pillow==8.0.1
pluggy==0.13.1
ply==3.11
portend==2.7.0
proglog==0.1.9
prompt-toolkit==3.0.18
punctuator==0.9.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.3.1
Pygments==2.8.0
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproj==3.0.1
pyquery==1.4.3
PySocks==1.7.1
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dev-tools==2020.9.10
python-docx==0.8.10
python-Levenshtein==0.12.1
pytube==11.0.0
pytube3==9.6.4
pytz==2020.5
pyupgrade==2.10.0
PyWavelets==1.1.1
pywebcopy==6.3.0
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
restructuredtext-lint==1.3.2
scipy==1.6.0
sgmllib3k==1.0.0
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
smart-open==4.1.0
smmap==3.0.5
snowballstemmer==2.1.0
snuggs==1.4.7
sortedcontainers==2.3.0
soupsieve==2.1
SpeechRecognition==3.8.1
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.26
sst==0.2.4
stevedore==3.3.0
subprocess.run==0.0.8
tempora==4.0.1
termcolor==1.1.0
testfixtures==6.17.1
Theano==1.0.5
tifffile==2021.3.31
tokenize-rt==4.1.0
toml==0.10.2
tox==3.22.0
tox-travis==0.12
tqdm==4.55.0
tube-dl==5.1.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.4.2
w3lib==1.22.0
wcwidth==0.2.5
websocket-client==0.48.0
wemake-python-styleguide==0.14.1
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2020.12.31
zc.lockfile==2.0
